+++ src/compiler/config/validate-config.ts
@@ -1,5 +1,5 @@
 import { createNodeLogger, createNodeSys } from '@sys-api-node';
-import { buildError, isBoolean, isNumber, isString, sortBy } from '@utils';
+import { buildError, buildWarn, isBoolean, isNumber, isString, sortBy } from '@utils';
 
 import {
   ConfigBundle,
@@ -182,6 +182,9 @@ export const validateConfig = (
     validatedConfig.extras.scopedSlotTextContentFix = !!validatedConfig.extras.scopedSlotTextContentFix;
   }
 
+  // TODO(STENCIL-1086): remove this option when it's the default behavior
+  validatedConfig.extras.experimentalScopedSlotChanges = !!validatedConfig.extras.experimentalScopedSlotChanges;
+
   setBooleanConfig(
     validatedConfig,
     'sourceMap',
@@ -261,6 +264,15 @@ export const validateConfig = (
     }
     return arr;
   }, [] as RegExp[]);
+
+  // TODO(STENCIL-1107): Remove this check. It'll be unneeded (and raise a compilation error when we build Stencil) once
+  // this property is removed.
+  if (validatedConfig.nodeResolve?.customResolveOptions) {
+    const warn = buildWarn(diagnostics);
+    // this message is particularly long - let the underlying logger implementation take responsibility for breaking it
+    // up to fit in a terminal window
+    warn.messageText = `nodeResolve.customResolveOptions is a deprecated option in a Stencil Configuration file. If you need this option, please open a new issue in the Stencil repository (https://github.com/ionic-team/stencil/issues/new/choose)`;
+  }
 
   CACHED_VALIDATED_CONFIG = validatedConfig;